+++ src/main/java/com/jtissdev/features/compta/entity/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package for compta entity.
+ */
+package com.jtissdev.features.compta.entity;