+++ src/main/java/com/jtissdev/features/pcg/controller/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package for pcg controller.
+ */
+package com.jtissdev.features.pcg.controller;