+++ src/main/java/com/jtissdev_API/engine/worker/package-info.java
@@ -1,4 +1,0 @@
-/**
- *
- */
-package com.jtissdev_API.engine.worker;