+++ src/main/java/com/jtissdev_API/features/PCG/service/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package for pcg service.
- */
-package com.jtissdev_API.features.pcg.service;