+++ src/main/java/com/jtissdev_API/features/compta/controller/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package for compta controller.
- */
-package com.jtissdev_API.features.compta.controller;