+++ src/main/java/com/jtissdev_API/features/compta/dto/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package for compta dto.
- */
-package com.jtissdev_API.features.compta.dto;