+++ src/main/java/com/jtissdev_API/features/compta/service/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package for compta service.
- */
-package com.jtissdev_API.features.compta.service;